Fix search for datatables
This patch fixes a bug which made it so that there was no
search field being added for datatables other than the main page.
As Nathan pointed out, this functionality broke in change 478423
because the dataTablesInit() had to check if #disabledatatables
was null. If #disabledatatables doesn't exist however, it should
still initialise the datatables.

Bug: T211781
Change-Id: If2d6efb54aba254cfb4594d25d781f266a784fcf