All notifications input checks all notification rows
The all notifications input in http://localhost:8000/my/preferences/ will now
check all notification rows via JS. I edited the toggleAllNotifs.js file that
is also used in watch.html, so that we can easily implement it elsewhere.
Only problem is that it will only work with one "all" form per page. But fixing
that is outside the scope of this task. (And also not yet neccessary)

Bug: T212033
Change-Id: I38a9ce3ab6b272da1dd045a87ea709acb6bbcdef