Create links from mentioned users
I'm afraid many users don't know that doing @username in a comment
will notify the user about the comment, so I decided to create a link
from that username to make it clear Tracker recognize this syntax.

As soon as users know Tracker recognizes this syntax, they might
think (correctly) that it emails the user and acts as a ping feature.

Change-Id: Icf681701f929b03669270d323cb3b66ee7aa8484