Add custom exception in get_content
With the current code it is very hard to understand the error
if a page doesn't exist and get_content try to get it. It isn't
fine explained because the error retuned is KeyError.
Among the approaches considered, this is the easier but useful way
to return a more useful
explanation of what happens: return a ValueError with a customized message.

A reviewer can review this change using the Django
shell (python manage.py shell) in trackersite directory.

Bug: T211663
Change-Id: Ia76827321d5c0f24f025afc7cacd47af5f6a75ed