Standardize user forms
Standardizes user forms so they all have a wide save/go button
at the bottom to make it easier to see the button. Currently, the
button looks like a input box instead of a button. This commit
also adds a standard Login/Register with MediaWiki to the Login
and register pages.

Bug: T240944
Change-Id: I900c1cfa2dc429405d403c374e17e97a99dc593a