commons-selection: Make it possible to filter either by username or filename
Previously, filename acted as a prefix, and since filtering was done
locally, it didn't offer any extra feature than Ctrl+F.

This changes interface feature a bit, but it should be more clear
what it is supposed to do.

Also console.error log when it is thrown, so there is
a way to discover what "unknown error" means.

Change-Id: I5207d42668ed6bbe23f6e0a6a70b0ac5f19472cd